add interpolation
fixed graph

+++ GraphProfilDialog.py
@@ -2551,7 +2551,7 @@ class GraphHydro(GraphCommon):
         # TODO
         if abs(val) == 10:
             var = 'selection'
-            val = val / 10
+            val = int(val / 10)
         else:
             var = self.inv
 
@@ -3081,7 +3081,7 @@ class GraphBasin(GraphCommon):
         # TODO
         if abs(val) == 10:
             # var = 'selection'
-            val = val / 10
+            val = int(val / 10)
         # else:
         #    var = self.inv
